fix: use in memory db in cloud env

+++ db/db.ts
@@ -1,4 +1,5 @@
 import { JSONFileSyncPreset } from "lowdb/node";
+import { MemorySync, LowSync } from "lowdb";
 import { BankAccountI, UserI, TransactionI } from "@/src/types";
 import fs from "node:fs";
 import path from "node:path";
@@ -9,6 +10,7 @@ type Data = {
   transactions: TransactionI[];
 };
 
+const IS_IN_CLOUD = process.env.IS_IN_CLOUD === "true";
 const DB_FILE_PATH = path.join(process.cwd(), "db.json");
 
 const defaultData: Data = {
@@ -17,14 +19,20 @@ const defaultData: Data = {
   transactions: [],
 };
 
+console.log("IS_IN_CLOUD", IS_IN_CLOUD);
 console.log("DB_FILE_PATH", DB_FILE_PATH);
 console.log("process.cwd()", process.cwd());
 
-if (!fs.existsSync(DB_FILE_PATH)) {
+if (!IS_IN_CLOUD && !fs.existsSync(DB_FILE_PATH)) {
   fs.writeFileSync(DB_FILE_PATH, JSON.stringify(defaultData));
 }
 
-const db = JSONFileSyncPreset(DB_FILE_PATH, defaultData);
+const db = IS_IN_CLOUD
+  ? new LowSync(new MemorySync<Data>(), {
+      ...defaultData,
+      user: { name: "Cloud User", id: "cloud-user" },
+    })
+  : JSONFileSyncPreset(DB_FILE_PATH, defaultData);
 
 db.read();
